stupid errors

+++ arch/geometry.h
@@ -2,17 +2,17 @@ void set_point(float x, float y) {
 	if (x > 83 || x < 0 || y > 47 || y < 0)
 		return;
 	int n = (uint16_t)(y / 8) * 84 + x;
-	screen_buffer[n] |= (uint8_t)pow(2, (uint8_t)y % 8);
+	screen_buffer[n] |= (uint8_t)pow(2, (uint16_t)y % 8);
 	lcd.setCursor(x, y/8);
 	lcd.send(HIGH, screen_buffer[n]);
 }
 
 void draw_line(char x1, char y1, char x2, char y2) {
-  uint8_t deltaX = abs(x2 - x1);
-  uint8_t deltaY = abs(y2 - y1);
+  uint16_t deltaX = abs(x2 - x1);
+  uint16_t deltaY = abs(y2 - y1);
   int8_t signX = x1 < x2 ? 1 : -1;
   int8_t signY = y1 < y2 ? 1 : -1;
-  int8_t error = (deltaX - deltaY) * 2;
+  int16_t error = (deltaX - deltaY) * 2;
   set_point(x2, y2);
 	while (pow(x1-x2, 2)+ pow(y1-y2, 2) > 1) {
 		set_point(x1, y1);
@@ -27,11 +27,11 @@ void draw_line(char x1, char y1, char x2, char y2) {
 	}
 }
 
-void draw_circle(int8_t X1, int8_t Y1, float R) {
-	int8_t x = 0;
-	int8_t y = R;
-	int8_t delta = 1 - 2 * R;
-	int8_t error = 0;
+void draw_circle(int16_t X1, int16_t Y1, float R) {
+	int16_t x = 0;
+	int16_t y = R;
+	int16_t delta = 1 - 2 * R;
+	int16_t error = 0;
 	while (y >= 0) {
 		set_point(X1 + x, Y1 + y);
 		set_point(X1 + x, Y1 - y);
